Add MIDI-CI protocol field to `fluid_synth_t`.
At this state we are unsure how the protocol setter is invoked, but in
either way it will be used to switch to new MIDI protocol.

MIDI-CI specifies that those CI messages are sent as Universal SysEx bytes
so it will be most likely that it can be changed at run time.